backend/requirements/dev.txt
-r base.txt
django-debug-toolbar>=4.2.0
pytest>=7.4.3
pytest-django>=4.7.0
